requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
mysqlclient==2.2.1
python-decouple==3.8
asgiref==3.6.0
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
mysqlclient==2.2.1
PyMySQL==1.1.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
